+++ src/mentee/mentee.module.ts
@@ -1,9 +1,9 @@
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { Mentee } from './entity/mentee.entity';
-import { MenteeInterest } from './entity/mentee-interest.entity';
-import { MenteeTechStack } from './entity/mentee-tech-stack.entity';
-import { MenteeService } from './mentee.service';
+import { Mentee } from './infrastructure/entity/mentee.entity';
+import { MenteeInterest } from './infrastructure/entity/mentee-interest.entity';
+import { MenteeTechStack } from './infrastructure/entity/mentee-tech-stack.entity';
+import { MenteeService } from './application/mentee.service';
 import { InterestModule } from '../interest/interest.module';
 
 @Module({
