+++ src/tech-stack/infrastructure/entity/tech-stack.entity.ts
@@ -1,7 +1,7 @@
 import { Column, Entity, OneToMany } from 'typeorm';
-import { BaseEntity } from '../../common/entity/base.entity';
-import { MentorTechStack } from '../../mentor/infrastructure/entity/mentor-tech-stack.entity';
-import { MenteeTechStack } from '../../mentee/entity/mentee-tech-stack.entity';
+import { BaseEntity } from '../../../common/entity/base.entity';
+import { MentorTechStack } from '../../../mentor/infrastructure/entity/mentor-tech-stack.entity';
+import { MenteeTechStack } from '../../../mentee/infrastructure/entity/mentee-tech-stack.entity';
 
 @Entity()
 export class TechStack extends BaseEntity {
